проблема отсутствия файла android_robot
