+++ app/build.gradle
@@ -8,7 +8,7 @@ android {
     buildToolsVersion "29.0.3"
 
     defaultConfig {
-        applicationId "ru.geekbrains.trainproj"
+        applicationId "ru.geekbrains.trainprojkotlin"
         minSdkVersion 22
         targetSdkVersion 29
         versionCode 1
@@ -42,9 +42,13 @@ dependencies {
 
     implementation 'com.google.firebase:firebase-core:17.5.0'
     implementation 'com.google.firebase:firebase-firestore:21.6.0'
+    implementation 'com.google.firebase:firebase-analytics:17.5.0'
+
 
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
+    implementation 'com.firebaseui:firebase-ui-auth:4.3.1'
+    implementation 'com.google.firebase:firebase-auth:19.4.0'
 }
 apply plugin: 'com.google.gms.google-services'
