--- app/src/main/java/ru/geekbrains/trainproj/data/model/NoteResult.kt
@@ -1,6 +1,0 @@
-package ru.geekbrains.trainproj.data.model
-
-sealed class NoteResult {
-    data class Success<out T>(val data: T) : NoteResult()
-    data class Error(val error: Throwable) : NoteResult()
-}
