--- app/src/main/java/ru/geekbrains/trainproj/data/provider/FirestoreProvider.kt
@@ -1,55 +1,0 @@
-package ru.geekbrains.trainproj.data.provider
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import com.google.firebase.firestore.FirebaseFirestore
-import ru.geekbrains.trainproj.data.entity.Note
-import ru.geekbrains.trainproj.data.model.NoteResult
-
-class FirestoreProvider : DataProvider {
-
-    companion object {
-        private const val NOTES_COLLECTION = "notes"
-    }
-
-    private val store = FirebaseFirestore.getInstance()
-    private val notesReference = store.collection(NOTES_COLLECTION)
-
-    override fun subscribeToAllNotes(): LiveData<NoteResult> {
-        val result = MutableLiveData<NoteResult>()
-
-        notesReference.addSnapshotListener { snapshot, e ->
-            e?.let {
-                result.value = NoteResult.Error(it)
-            } ?: snapshot?.let {
-                val notes = snapshot.documents.mapNotNull { it.toObject(Note::class.java) }
-                result.value = NoteResult.Success(notes)
-            }
-        }
-
-        return result
-    }
-
-    override fun saveNote(note: Note): LiveData<NoteResult> {
-        val result = MutableLiveData<NoteResult>()
-
-        notesReference.document(note.id).set(note).addOnSuccessListener { snapshot ->
-            result.value = NoteResult.Success(note)
-        }.addOnFailureListener() {
-            result.value = NoteResult.Error(it)
-        }
-        return result
-    }
-
-    override fun getNoteById(id: String): LiveData<NoteResult> {
-        val result = MutableLiveData<NoteResult>()
-
-        notesReference.document(id).get().addOnSuccessListener { snapshot ->
-            val note = snapshot.toObject(Note::class.java)
-            result.value = NoteResult.Success(note)
-        }.addOnFailureListener {
-            result.value = NoteResult.Error(it)
-        }
-        return result
-    }
-}
