--- app/src/main/java/ru/geekbrains/trainproj/ui/base/BaseActivity.kt
@@ -1,37 +1,0 @@
-package ru.geekbrains.trainproj.ui.base
-
-import android.os.Bundle
-import android.os.PersistableBundle
-import android.widget.Toast
-import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.Observer
-
-abstract class BaseActivity<T, S : BaseViewState<T>> : AppCompatActivity() {
-
-    abstract val viewModel: BaseViewModel<T, S>
-    abstract val layoutRes: Int
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        viewModel.getViewState().observe(this, Observer { state ->
-            state ?: return@Observer
-            state.error?.let {
-                renderError(it)
-                return@Observer
-            }
-            renderData(state.data)
-        })
-    }
-
-    abstract fun renderData(data: T)
-
-    protected fun renderError(error: Throwable) {
-        error.message?.let {
-            showError(it)
-        }
-    }
-
-    protected fun showError(error: String) {
-        Toast.makeText(this, error, Toast.LENGTH_SHORT).show()
-    }
-}
