--- app/src/main/java/ru/geekbrains/trainproj/ui/base/BaseViewModel.kt
@@ -1,12 +1,0 @@
-package ru.geekbrains.trainproj.ui.base
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-open class BaseViewModel <T, S: BaseViewState<T>>: ViewModel(){
-    open val viewStateLiveData = MutableLiveData<S>()
-    open fun getViewState(): LiveData<S> = viewStateLiveData
-}
-
-
