--- app/src/main/java/ru/geekbrains/trainproj/ui/base/BaseViewState.kt
@@ -1,3 +1,0 @@
-package ru.geekbrains.trainproj.ui.base
-
-open class BaseViewState <T>(val data: T, val error: Throwable?)
