--- app/src/main/java/ru/geekbrains/trainproj/ui/main/MainActivity.kt
@@ -1,45 +1,0 @@
-package ru.geekbrains.trainproj.ui.main
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.widget.Toast
-import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.ViewModelProviders
-import androidx.recyclerview.widget.GridLayoutManager
-import kotlinx.android.synthetic.main.activity_main.*
-import ru.geekbrains.trainproj.R
-import ru.geekbrains.trainproj.data.entity.Note
-import ru.geekbrains.trainproj.ui.base.BaseActivity
-import ru.geekbrains.trainproj.ui.main.MainViewModel
-import ru.geekbrains.trainproj.ui.main.NotesRVAdapter
-import ru.geekbrains.trainproj.ui.note.NoteActivity
-
-class MainActivity : BaseActivity<List<Note>?, MainViewState>() {
-
-    override val viewModel: MainViewModel by lazy {
-        ViewModelProviders.of(this).get(MainViewModel::class.java)
-    }
-    lateinit var adapter: NotesRVAdapter
-    override val layoutRes: Int = R.layout.activity_main
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        rv_notes.layoutManager = GridLayoutManager(this, 2)
-        adapter = NotesRVAdapter{
-            NoteActivity.start(this, it.id)
-        }
-        rv_notes.adapter = adapter
-
-        fab.setOnClickListener {
-            NoteActivity.start(this)
-        }
-    }
-
-    override fun renderData(data: List<Note>?) {
-        data?.let {
-            adapter.notes = it
-        }
-    }
-}
