--- app/src/main/java/ru/geekbrains/trainproj/ui/main/Model.kt
@@ -1,17 +1,0 @@
-package ru.geekbrains.trainproj.ui.main
-
-import androidx.lifecycle.MutableLiveData
-
-class Model {
-    private var counter = 0
-    private val counterLiveData = MutableLiveData<Int>()
-
-    fun counterLiveData() :MutableLiveData<Int> = counterLiveData
-
-    fun riseCounter() {
-        counter++
-        counterLiveData.value = counter
-    }
-}
-
-
