+++ app/src/main/java/ru/geekbrains/trainprojkotlin/data/entity/User.kt
@@ -1,0 +1,3 @@
+package ru.geekbrains.trainprojkotlin.data.entity
+
+class User(val name : String, val email : String)
