+++ app/src/main/java/ru/geekbrains/trainprojkotlin/data/errors/NoAuthException.kt
@@ -1,0 +1,6 @@
+package ru.geekbrains.trainprojkotlin.data.errors
+
+
+class NoAuthException(msg: String? = null): Throwable(msg) {
+
+}
