+++ app/src/main/java/ru/geekbrains/trainprojkotlin/data/model/NoteResult.kt
@@ -1,0 +1,6 @@
+package ru.geekbrains.trainprojkotlin.data.model
+
+sealed class NoteResult {
+    data class Success<out T>(val data: T) : NoteResult()
+    data class Error(val error: Throwable) : NoteResult()
+}
