+++ app/src/main/java/ru/geekbrains/trainprojkotlin/ui/base/BaseViewModel.kt
@@ -1,0 +1,12 @@
+package ru.geekbrains.trainprojkotlin.ui.base
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+open class BaseViewModel <T, S: BaseViewState<T>>: ViewModel(){
+    open val viewStateLiveData = MutableLiveData<S>()
+    open fun getViewState(): LiveData<S> = viewStateLiveData
+}
+
+
