+++ app/src/main/java/ru/geekbrains/trainprojkotlin/ui/base/BaseViewState.kt
@@ -1,0 +1,3 @@
+package ru.geekbrains.trainprojkotlin.ui.base
+
+open class BaseViewState <T>(val data: T, val error: Throwable?)
