+++ app/src/main/java/ru/geekbrains/trainprojkotlin/ui/main/LogoutDialog.kt
@@ -1,0 +1,26 @@
+package ru.geekbrains.trainprojkotlin.ui.main
+
+import android.app.AlertDialog
+import android.app.Dialog
+import android.os.Bundle
+import androidx.fragment.app.DialogFragment
+import ru.geekbrains.trainprojkotlin.R
+
+class LogoutDialog : DialogFragment() {
+    companion object {
+        val TAG = LogoutDialog::class.java.name + "TAG"
+        fun createInstance() = LogoutDialog()
+    }
+
+    override fun onCreateDialog(savedInstanceState: Bundle?): Dialog =
+        AlertDialog.Builder(context!!)
+            .setTitle(R.string.logout_dialog_title)
+            .setMessage(R.string.logout_dialog_message)
+            .setPositiveButton(R.string.ok_bth_title) { _, _ ->  (activity as LogoutListener).onLogout() }
+            .setNegativeButton(R.string.logout_dialog_cancel) {_, _ -> dismiss() }
+            .create()
+
+    interface LogoutListener {
+        fun onLogout()
+    }
+}
