+++ app/src/main/java/ru/geekbrains/trainprojkotlin/ui/main/MainActivity.kt
@@ -1,0 +1,77 @@
+package ru.geekbrains.trainprojkotlin.ui.main
+
+import android.content.Context
+import android.content.Intent
+import android.os.Bundle
+import android.view.Menu
+import android.view.MenuInflater
+import android.view.MenuItem
+import androidx.lifecycle.ViewModelProviders
+import androidx.recyclerview.widget.GridLayoutManager
+import com.firebase.ui.auth.AuthUI
+import kotlinx.android.synthetic.main.activity_main.*
+import ru.geekbrains.trainprojkotlin.R
+import ru.geekbrains.trainprojkotlin.data.entity.Note
+import ru.geekbrains.trainprojkotlin.ui.base.BaseActivity
+import ru.geekbrains.trainprojkotlin.ui.note.NoteActivity
+import ru.geekbrains.trainprojkotlin.ui.splash.SplashActivity
+
+
+class MainActivity : BaseActivity<List<Note>?, MainViewState>(), LogoutDialog.LogoutListener {
+
+    override val viewModel: MainViewModel by lazy {
+        ViewModelProviders.of(this).get(MainViewModel::class.java)
+    }
+    lateinit var adapter: NotesRVAdapter
+    override val layoutRes: Int = R.layout.activity_main
+
+    companion object {
+        fun getStartIntent(context: Context) = Intent(context,
+            MainActivity::class.java)
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setSupportActionBar(toolbar)
+
+        rv_notes.layoutManager = GridLayoutManager(this, 2)
+        adapter = NotesRVAdapter {
+            NoteActivity.start(this, it.id)
+        }
+        rv_notes.adapter = adapter
+
+        fab.setOnClickListener {
+            NoteActivity.start(this)
+        }
+    }
+
+    override fun renderData(data: List<Note>?) {
+        data?.let {
+            adapter.notes = it
+        }
+    }
+
+    override fun onCreateOptionsMenu(menu: Menu?): Boolean =
+        MenuInflater(this).inflate(R.menu.main, menu).let { true }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean =
+        when (item.itemId) {
+            R.id.logout -> showLogoutDialog().let { true }
+            else -> false
+        }
+
+    fun showLogoutDialog() {
+        supportFragmentManager.findFragmentByTag(LogoutDialog.TAG) ?: LogoutDialog().show(
+            supportFragmentManager,
+            LogoutDialog.TAG)
+    }
+
+    override fun onLogout() {
+        AuthUI.getInstance()
+            .signOut(this)
+            .addOnCompleteListener {
+                startActivity(Intent(this, SplashActivity::class.java))
+                finish()
+            }
+    }
+}
