+++ app/src/main/java/ru/geekbrains/trainprojkotlin/ui/main/MainViewModel.kt
@@ -1,0 +1,32 @@
+package ru.geekbrains.trainprojkotlin.ui.main
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModel
+import ru.geekbrains.trainprojkotlin.data.NotesRepository
+import ru.geekbrains.trainprojkotlin.data.entity.Note
+import ru.geekbrains.trainprojkotlin.data.model.NoteResult
+import ru.geekbrains.trainprojkotlin.ui.base.BaseViewModel
+
+class MainViewModel : BaseViewModel<List<Note>?, MainViewState>() {
+
+    private val notesObserver = Observer<NoteResult>{result ->
+        result ?: return@Observer
+        when(result){
+            is NoteResult.Success<*> -> viewStateLiveData.value = MainViewState(notes = result.data as? List<Note>)
+            is NoteResult.Error -> viewStateLiveData.value = MainViewState(error = result.error)
+        }
+    }
+
+  private val repositoryNotes = NotesRepository.getNotes()
+    init {
+        viewStateLiveData.value = MainViewState()
+        repositoryNotes.observeForever(notesObserver)
+        }
+
+    override fun onCleared() {
+        super.onCleared()
+        repositoryNotes.removeObserver(notesObserver)
+    }
+}
