+++ app/src/main/java/ru/geekbrains/trainprojkotlin/ui/main/Model.kt
@@ -1,0 +1,17 @@
+package ru.geekbrains.trainprojkotlin.ui.main
+
+import androidx.lifecycle.MutableLiveData
+
+class Model {
+    private var counter = 0
+    private val counterLiveData = MutableLiveData<Int>()
+
+    fun counterLiveData() :MutableLiveData<Int> = counterLiveData
+
+    fun riseCounter() {
+        counter++
+        counterLiveData.value = counter
+    }
+}
+
+
