+++ app/src/main/java/ru/geekbrains/trainprojkotlin/ui/note/NoteActivity.kt
@@ -1,0 +1,112 @@
+package ru.geekbrains.trainprojkotlin.ui.note
+
+import android.content.Context
+import android.content.Intent
+import android.os.Bundle
+import android.text.Editable
+import android.text.TextWatcher
+import android.view.MenuItem
+import androidx.core.content.res.ResourcesCompat
+import androidx.lifecycle.ViewModelProviders
+import kotlinx.android.synthetic.main.activity_main.toolbar
+import kotlinx.android.synthetic.main.activity_note.*
+import ru.geekbrains.trainprojkotlin.R
+import ru.geekbrains.trainprojkotlin.data.entity.Note
+import ru.geekbrains.trainprojkotlin.ui.base.BaseActivity
+import java.text.SimpleDateFormat
+import java.util.*
+
+class NoteActivity : BaseActivity<Note?, NoteViewState>() {
+    companion object {
+        private const val NOTE_KEY = "note"
+        private const val DATE_FORMAT = "dd.MM.yy HH.mm"
+
+        fun start(context: Context, noteId: String? = null) =
+            Intent(context, NoteActivity::class.java).apply {
+                putExtra(NOTE_KEY, noteId)
+                context.startActivity(this)
+            }
+    }
+
+    private var note: Note? = null
+    override val viewModel by lazy {
+        ViewModelProviders.of(this).get(NoteViewModel::class.java)
+    }
+
+    val textWatcher = object : TextWatcher {
+        override fun afterTextChanged(s: Editable?) {
+            saveNote()
+        }
+
+        override fun beforeTextChanged(s: CharSequence?, start: Int, count: Int, after: Int) {}
+        override fun onTextChanged(s: CharSequence?, start: Int, before: Int, count: Int) {}
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_note)
+        setSupportActionBar(toolbar)
+        supportActionBar?.setDisplayHomeAsUpEnabled(true)
+
+        val noteId = intent.getStringExtra(NOTE_KEY)
+
+        noteId?.let {
+            viewModel.loadNote(it)
+        } ?: let {
+            supportActionBar?.title = getString(R.string.new_note)
+            initView()
+        }
+    }
+
+    override fun renderData(data: Note?) {
+        supportActionBar?.title = note?.let {
+            SimpleDateFormat(DATE_FORMAT, Locale.getDefault()).format(it.lastChanged)
+        } ?: getString(R.string.new_note)
+        initView()
+    }
+
+    private fun initView() {
+        et_title.removeTextChangedListener(textWatcher)
+        et_body.removeTextChangedListener(textWatcher)
+
+        note?.let {
+            et_title.setTextKeepState(it.title)
+            et_body.setTextKeepState(it.text)
+            val color = when (it.color) {
+                Note.Color.WHITE -> R.color.white
+                Note.Color.VIOLET -> R.color.violet
+                Note.Color.YELLOW -> R.color.yellow
+                Note.Color.RED -> R.color.red
+                Note.Color.PINK -> R.color.pink
+                Note.Color.GREEN -> R.color.green
+            }
+            toolbar.setBackgroundColor(ResourcesCompat.getColor(resources, color, null))
+        }
+        et_title.addTextChangedListener(textWatcher)
+        et_body.addTextChangedListener(textWatcher)
+    }
+
+    private fun saveNote() {
+        et_title.text?.let {
+            if (it.length < 3) return
+        } ?: return
+
+        note = note?.copy(
+            title = et_title.text.toString(),
+            text = et_body.text.toString(),
+            lastChanged = Date()
+        ) ?: Note(UUID.randomUUID().toString(), et_title.text.toString(), et_body.text.toString())
+
+        note?.let { viewModel.save(it) }
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem) = when (item.itemId) {
+        android.R.id.home -> {
+            onBackPressed()
+            true
+        }
+        else -> super.onOptionsItemSelected(item)
+    }
+
+    override val layoutRes: Int = R.layout.activity_note
+}
