+++ app/src/main/java/ru/geekbrains/trainprojkotlin/ui/splash/SplashActivity.kt
@@ -1,0 +1,33 @@
+package ru.geekbrains.trainprojkotlin.ui.splash
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.ViewModelProviders
+import ru.geekbrains.trainprojkotlin.ui.base.BaseActivity
+
+import ru.geekbrains.trainprojkotlin.ui.main.MainActivity
+
+class SplashActivity : BaseActivity<Boolean?, SplashViewState>() {
+    override val viewModel: SplashViewModel by lazy {
+        ViewModelProviders.of(this).get(SplashViewModel::class.java)
+    }
+
+    override val layoutRes = null
+
+    override fun onResume() {
+        super.onResume()
+        viewModel.requestUser()
+    }
+
+    override fun renderData(data: Boolean?) {
+        data?.takeIf { it }?.let {
+            startMainActivity()
+        }
+    }
+
+    fun startMainActivity() {
+        startActivity(MainActivity.getStartIntent(this))
+        finish()
+    }
+
+}
