+++ app/src/main/java/ru/geekbrains/trainprojkotlin/ui/splash/SplashViewState.kt
@@ -1,0 +1,5 @@
+package ru.geekbrains.trainprojkotlin.ui.splash
+
+import ru.geekbrains.trainprojkotlin.ui.base.BaseViewState
+
+class SplashViewState(authenticated: Boolean? = null, error: Throwable? = null): BaseViewState<Boolean?>(authenticated, error)
