+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,9 @@
 <resources>
     <string name="app_name">TrainProj</string>
     <string name="new_note">Новая заметка</string>
+    <string name="logout_menu_title">Выход</string>
+    <string name="logout_dialog_title">Выход</string>
+    <string name="logout_dialog_message">Вы уверены?</string>
+    <string name="ok_bth_title">Да</string>
+    <string name="logout_dialog_cancel">Нет</string>
 </resources>
